src/main/java/com/mobiusk/vrsvp/command/SlashCommandReply.java
package com.mobiusk.vrsvp.command;

import com.mobiusk.vrsvp.button.ButtonUi;
import com.mobiusk.vrsvp.embed.EmbedUi;
import lombok.RequiredArgsConstructor;
import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;

import javax.annotation.Nonnull;

@RequiredArgsConstructor
public class SlashCommandReply {

	// Class constructor field(s)
	private final ButtonUi buttonUi;
	private final EmbedUi embedUi;

	/**
	 * Build an RSVP form everyone can see with a given number of time slot, each with an index and incremented timestamp title.
	 */
	public void rsvpCreation(@Nonnull SlashCommandInteractionEvent event, @Nonnull SlashCommandInputs inputs) {
		event.reply("")
			.addEmbeds(embedUi.build(inputs))
			.addActionRow(buttonUi.buildRsvpActionPrompts())
			.queue();
	}

}
